requirementas.txt
torch==1.12.0
streamlit==1.41.1
pillow==10.4.0
torchvision==0.13.0